+++ packages/preview-server/src/actions/get-templates-directory-metadata-action.ts
@@ -1,7 +1,9 @@
 'use server';
 
-import type { TemplatesDirectory } from '../utils/get-templates-directory-metadata';
-import { getTemplatesDirectoryMetadata } from '../utils/get-templates-directory-metadata';
+import {
+  getTemplatesDirectoryMetadata,
+  type TemplatesDirectory,
+} from '@ahmedrowaihi/pdf-forge-templates';
 
 export const getTemplatesDirectoryMetadataAction = async (
   absolutePathToTemplatesDirectory: string,